Moved and shortened snippets
