+++ slick/src/test/java/docs/javadsl/DocSnippetFlow.java
@@ -2,12 +2,10 @@
  * Copyright (C) 2016-2018 Lightbend Inc. <http://www.lightbend.com>
  */
 
-package example;
+package docs.javadsl;
 
-// #sink-example
 import java.util.List;
 import java.util.concurrent.CompletionStage;
-import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
@@ -19,12 +17,14 @@ import akka.stream.Materializer;
 import akka.stream.javadsl.*;
 import akka.stream.alpakka.slick.javadsl.*;
 
-public class DocSnippetSink {
+public class DocSnippetFlow {
   public static void main(String[] args) throws Exception {
     final ActorSystem system = ActorSystem.create();
     final Materializer materializer = ActorMaterializer.create(system);
 
+    // #flow-example
     final SlickSession session = SlickSession.forConfig("slick-h2");
+    system.registerOnTermination(session::close);
 
     final List<User> users =
         IntStream.range(0, 42)
@@ -34,23 +34,23 @@ public class DocSnippetSink {
 
     final CompletionStage<Done> done =
         Source.from(users)
-            .runWith(
-                Slick.<User>sink(
+            .via(
+                Slick.<User>flow(
                     session,
-                    // add an optional second argument to specify the parallism factor (int)
+                    // add an optional second argument to specify the parallism facsltor (int)
                     (user) ->
                         "INSERT INTO ALPAKKA_SLICK_JAVADSL_TEST_USERS VALUES ("
                             + user.id
                             + ", '"
                             + user.name
-                            + "')"),
-                materializer);
+                            + "')"))
+            .log("nr-of-updated-rows")
+            .runWith(Sink.ignore(), materializer);
+    // #flow-example
 
     done.whenComplete(
         (value, exception) -> {
-          session.close();
           system.terminate();
         });
   }
 }
-// #sink-example
