+++ slick/src/test/java/docs/javadsl/User.java
@@ -2,7 +2,7 @@
  * Copyright (C) 2016-2018 Lightbend Inc. <http://www.lightbend.com>
  */
 
-package akka.stream.alpakka.slick.javadsl;
+package docs.javadsl;
 
 public class User {
   public final Integer id;
